# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"PostResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"hasPinned": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"caption",
					"createdAt",
					"userId",
					"username",
					"avatarUrl",
					"hasPinned"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostBase64Input": {
				"properties": {
					"imageBase64": {
						"type": "string"
					},
					"imageFileType": {
						"type": "string"
					},
					"caption": {
						"type": "string"
					}
				},
				"required": [
					"imageBase64",
					"imageFileType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommentResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"text": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"postId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"text",
					"userId",
					"postId",
					"username",
					"avatarUrl",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentInput": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Pinterest Simplified API",
		"version": "1.0.0",
		"description": "API docs for simplified Pinterest backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Your Name"
		}
	},
	"paths": {
		"/posts": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Post Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostBase64Input"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFeedPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PostResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/search": {
			"get": {
				"operationId": "SearchPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PostResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}": {
			"get": {
				"operationId": "GetPostById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Posts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/users/{userId}/posts": {
			"get": {
				"operationId": "GetPostsByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PostResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/posts/{postId}/pin": {
			"post": {
				"operationId": "PinPost",
				"responses": {
					"201": {
						"description": "Liked",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Pin & Comments)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}/unpin": {
			"delete": {
				"operationId": "UnpinPost",
				"responses": {
					"200": {
						"description": "Unpinned",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Pin & Comments)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/posts/{postId}/comments": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"201": {
						"description": "Comment Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Pin & Comments)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CommentResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Interactions (Pin & Comments)"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}